fix: hook not called inside plate editor
